+++ .gitignore
@@ -1,0 +1,3 @@
+*.sw[op]
+config.json
+pkg-build
