+++ config.template.json
@@ -1,0 +1,6 @@
+{
+    "namecom": {
+      "username": "you@your-domain.com",
+      "token": "abcde"
+    }
+}
