+++ pkg-src/deb/certbot-dns-namecom/DEBIAN/postrm
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+if [ "$1" == "purge" ]; then
+    if [ -e /usr/share/debconf/confmodule ]; then
+        . /usr/share/debconf/confmodule
+        db_purge
+    fi
+fi
