+++ pkg-src/deb/certbot-dns-namecom/DEBIAN/prerm
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+if [ "$1" == 'purge' ]; then
+  CONFFILE=/etc/certbot-dns-namecom.config.json
+  rm "$CONFFILE"
+fi
